calendar weeks are now computed, the logic is fixed and the code is cleaner
